scripts/requirements-recipe-pdf.txt
markdown>=3.5
weasyprint>=62.0
PyYAML>=6.0
